# contracts/generated/model-api-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "BriefRequest": {
        "additionalProperties": false,
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/IncidentAnalyzeRequest",
            "description": "기존 분석 요청. 전사 원문은 수정하지 않습니다."
          },
          "invalidated_confirmation_ids": {
            "description": "이번 입력에서 취소된 확인 ID. 해당 역할의 확인을 제거합니다.",
            "items": {
              "type": "string"
            },
            "maxItems": 2,
            "title": "Invalidated Confirmation Ids",
            "type": "array"
          },
          "reported_evidence_conflict": {
            "default": false,
            "description": "새 근거 상충을 보고하면 양쪽 확인을 보류합니다. false는 상충 없음의 증명이 아닙니다.",
            "title": "Reported Evidence Conflict",
            "type": "boolean"
          },
          "revision": {
            "description": "Backend가 관리하는 사고 revision. 모델 서버가 최신 여부를 보증하지 않습니다.",
            "minimum": 0.0,
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "analysis",
          "revision"
        ],
        "title": "BriefRequest",
        "type": "object"
      },
...
      "ConfirmationBasis": {
        "enum": [
          "CONTAINER_LABEL",
          "SITE_MSDS",
          "SHIPPING_DOCUMENT",
          "INSTRUMENT_READING",
          "RESPONDER_OBSERVATION",
          "OTHER_VERIFIED_SOURCE"
        ],
        "title": "ConfirmationBasis",
        "type": "string"
      },
...
      "ConfirmedSubstanceInput": {
        "additionalProperties": false,
        "description": "백엔드가 인증된 대원 확인 레코드에서 전달하는 물질.\n\n``confirmation_id``는 사용자 자유입력 ID가 아니라 인증된 백엔드가 보관한\n현장 확인 레코드 식별자여야 한다. 모델 API는 이 레코드 참조와 확인 시각을\n필수로 받으며, 인증되지 않은 후보 입력과 타입 수준에서 분리한다.",
        "properties": {
          "cas_number": {
            "maxLength": 12,
            "minLength": 5,
            "title": "Cas Number",
            "type": "string"
          },
          "confirmation_basis": {
            "$ref": "#/components/schemas/ConfirmationBasis"
          },
          "confirmation_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Confirmation Id",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "presence_status": {
            "const": "CONFIRMED_PRESENT",
            "title": "Presence Status",
            "type": "string"
          },
          "role": {
            "enum": [
              "INCIDENT",
              "FACILITY"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "confirmation_id",
          "cas_number",
          "role",
          "presence_status",
          "confirmation_basis",
          "observed_at"
        ],
        "title": "ConfirmedSubstanceInput",
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fields": {
            "items": {
              "type": "string"
            },
            "title": "Fields",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "occurred_at_utc": {
            "format": "date-time",
            "title": "Occurred At Utc",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "schema_version": {
            "const": "chemiguard119-api-v1",
            "default": "chemiguard119-api-v1",
            "title": "Schema Version",
            "type": "string"
          },
          "service_name": {
            "const": "케미체크119",
            "default": "케미체크119",
            "title": "Service Name",
            "type": "string"
          }
        },
        "required": [
          "error",
          "request_id",
          "occurred_at_utc"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "IncidentAnalyzeRequest": {
        "additionalProperties": false,
        "examples": [
          {
            "evidence_top_k": 5,
            "incident_id": "INC-20260721-0001",
            "input": {
              "occurred_at": "2026-07-21T19:20:00+09:00",
              "text": "차아염소산나트륨 탱크가 누출됐고 옆 저장고에는 염산이 있습니다.",
              "type": "VOICE_TRANSCRIPT"
            },
            "location": {
              "address": "경기 화성시 팔탄면",
              "coordinate_source": "DISPATCH_SYSTEM",
              "facility_name": "OO전자 공장",
              "latitude": 37.2181,
              "longitude": 126.9417,
              "province": "경기도",
              "resolved_at": "2026-07-21T19:20:00+09:00"
            },
            "operations_context": {
              "dispatch_station_name": "화성소방서",
              "journey_state": "EN_ROUTE",
              "responder_position": {
                "accuracy_m": 12.0,
                "latitude": 37.2065,
                "longitude": 126.8311,
                "observed_at": "2026-07-21T19:22:00+09:00",
                "source": "MDT_DEVICE_GPS"
              }
            },
            "planned_actions": [
              {
                "raw_text": "누출구역 통제"
              }
            ],
            "request_id": "REQ-20260721-0001"
          }
        ],
        "properties": {
          "confirmed_facility_substance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfirmedSubstanceInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "confirmed_incident_substance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfirmedSubstanceInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence_top_k": {
            "default": 5,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Evidence Top K",
            "type": "integer"
          },
          "incident_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_.:-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Id"
          },
          "input": {
            "$ref": "#/components/schemas/IncidentInput"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IncidentLocation"
              },
              {
                "type": "null"
              }
            ]
          },
          "operations_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OperationsContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "planned_actions": {
            "items": {
              "$ref": "#/components/schemas/PlannedActionInput"
            },
            "maxItems": 20,
            "title": "Planned Actions",
            "type": "array"
          },
          "request_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_.:-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "input"
        ],
        "title": "IncidentAnalyzeRequest",
        "type": "object"
      },
      "IncidentInput": {
        "additionalProperties": false,
        "properties": {
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "text": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/IncidentInputType",
            "default": "MANUAL_TEXT"
          }
        },
        "required": [
          "text"
        ],
        "title": "IncidentInput",
        "type": "object"
      },
      "IncidentInputType": {
        "enum": [
          "MANUAL_TEXT",
          "DISPATCH_TEXT",
          "VOICE_TRANSCRIPT",
          "STRUCTURED_FORM"
        ],
        "title": "IncidentInputType",
        "type": "string"
      },
      "IncidentLocation": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "coordinate_source": {
            "anyOf": [
              {
                "enum": [
                  "DISPATCH_SYSTEM",
                  "GEOCODING_PROVIDER",
                  "RESPONDER_OBSERVATION",
                  "MANUAL_ENTRY",
                  "DEMO_FIXTURE"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coordinate Source"
          },
          "facility_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facility Name"
          },
          "geocoding_provider": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geocoding Provider"
          },
          "latitude": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "province": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Province"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          }
        },
        "title": "IncidentLocation",
        "type": "object"
      },
...
      "OperationsContext": {
        "additionalProperties": false,
        "properties": {
          "dispatch_station_name": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispatch Station Name"
          },
          "journey_state": {
            "default": "DISPATCHED",
            "enum": [
              "DISPATCHED",
              "EN_ROUTE",
              "ARRIVED",
              "ON_SCENE"
            ],
            "title": "Journey State",
            "type": "string"
          },
          "responder_position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponderPosition"
              },
              {
                "type": "null"
              }
            ]
          },
          "route": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServerRoute"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "OperationsContext",
        "type": "object"
      },
      "PlannedActionInput": {
        "additionalProperties": false,
        "properties": {
          "raw_text": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Raw Text",
            "type": "string"
          }
        },
        "required": [
          "raw_text"
        ],
        "title": "PlannedActionInput",
        "type": "object"
      },
...
      "ResponderPosition": {
        "additionalProperties": false,
        "properties": {
          "accuracy_m": {
            "anyOf": [
              {
                "maximum": 5000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy M"
          },
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "source": {
            "enum": [
              "VEHICLE_GPS",
              "MDT_DEVICE_GPS",
              "MANUAL_DISPATCH",
              "DEMO_SIMULATION"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "latitude",
          "longitude",
          "observed_at",
          "source"
        ],
        "title": "ResponderPosition",
        "type": "object"
      },
      "RouteGeometry": {
        "additionalProperties": false,
        "properties": {
          "coordinates": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "maxItems": 10000,
            "minItems": 2,
            "title": "Coordinates",
            "type": "array"
          },
          "type": {
            "const": "LineString",
            "default": "LineString",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "coordinates"
        ],
        "title": "RouteGeometry",
        "type": "object"
      },
      "ServerRoute": {
        "additionalProperties": false,
        "description": "BE가 비밀키로 호출한 길찾기 사업자의 결과.\n\n모델 API는 이 값을 재계산하거나 보정하지 않는다. 데모 경로는 반드시\n``DEMO_SIMULATION``으로 표시해 실제 길찾기 결과와 혼동되지 않게 한다.",
        "properties": {
          "attribution": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Attribution",
            "type": "string"
          },
          "distance_m": {
            "exclusiveMinimum": 0.0,
            "maximum": 5000000.0,
            "title": "Distance M",
            "type": "integer"
          },
          "duration_seconds": {
            "exclusiveMinimum": 0.0,
            "maximum": 604800.0,
            "title": "Duration Seconds",
            "type": "integer"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "geometry": {
            "$ref": "#/components/schemas/RouteGeometry"
          },
          "mode": {
            "enum": [
              "LIVE_API",
              "CACHED_API",
              "DEMO_SIMULATION"
            ],
            "title": "Mode",
            "type": "string"
          },
          "provider": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Provider",
            "type": "string"
          },
          "provider_reference": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Reference"
          },
          "remaining_distance_m": {
            "maximum": 5000000.0,
            "minimum": 0.0,
            "title": "Remaining Distance M",
            "type": "integer"
          },
          "remaining_duration_seconds": {
            "maximum": 604800.0,
            "minimum": 0.0,
            "title": "Remaining Duration Seconds",
            "type": "integer"
          },
          "route_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Route Id",
            "type": "string"
          },
          "traffic_applied": {
            "title": "Traffic Applied",
            "type": "boolean"
          }
        },
        "required": [
          "provider",
          "mode",
          "route_id",
          "geometry",
          "distance_m",
          "duration_seconds",
          "remaining_distance_m",
          "remaining_duration_seconds",
          "generated_at",
          "traffic_applied",
          "attribution"
        ],
        "title": "ServerRoute",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "Backend 전용 API 키. 로컬 데모 키를 운영에 사용하지 마세요.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "화학사고 신고 구조화, 물질 후보 검색, 공식 근거 검색, 근거 제한형 RAG와 공개근거 CAMEO 스크리닝 또는 승인 Rule 조회를 제공하는 의사결정 지원 API입니다. 모델 출력은 현장 명령이 아닙니다.",
    "title": "케미체크119 모델 API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/agents/incidents/brief/stream": {
      "post": {
        "operationId": "incident_brief_stream_api_v1_agents_incidents_brief_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "both_confirmed": {
                  "summary": "합성 예시: 두 역할 확인",
                  "value": {
                    "analysis": {
                      "confirmed_facility_substance": {
                        "cas_number": "7647-01-0",
                        "confirmation_basis": "CONTAINER_LABEL",
                        "confirmation_id": "CNF-DEMO-FACILITY",
                        "observed_at": "2026-09-01T00:00:00+09:00",
                        "presence_status": "CONFIRMED_PRESENT",
                        "role": "FACILITY"
                      },
                      "confirmed_incident_substance": {
                        "cas_number": "7681-52-9",
                        "confirmation_basis": "CONTAINER_LABEL",
                        "confirmation_id": "CNF-DEMO-INCIDENT",
                        "observed_at": "2026-09-01T00:00:00+09:00",
                        "presence_status": "CONFIRMED_PRESENT",
                        "role": "INCIDENT"
                      },
                      "evidence_top_k": 5,
                      "incident_id": "INC-SYNTHETIC-1",
                      "input": {
                        "text": "차아염소산나트륨 탱크에서 누출이 있고 옆 저장고에는 염산이 있습니다.",
                        "type": "VOICE_TRANSCRIPT"
                      },
                      "request_id": "REQ-DEMO-1"
                    },
                    "revision": 3
                  }
                },
                "cancelled": {
                  "summary": "합성 예시: 확인 취소 후 재분석",
                  "value": {
                    "analysis": {
                      "confirmed_facility_substance": {
                        "cas_number": "7647-01-0",
                        "confirmation_basis": "CONTAINER_LABEL",
                        "confirmation_id": "CNF-DEMO-FACILITY",
                        "observed_at": "2026-09-01T00:00:00+09:00",
                        "presence_status": "CONFIRMED_PRESENT",
                        "role": "FACILITY"
                      },
                      "confirmed_incident_substance": {
                        "cas_number": "7681-52-9",
                        "confirmation_basis": "CONTAINER_LABEL",
                        "confirmation_id": "CNF-DEMO-INCIDENT",
                        "observed_at": "2026-09-01T00:00:00+09:00",
                        "presence_status": "CONFIRMED_PRESENT",
                        "role": "INCIDENT"
                      },
                      "evidence_top_k": 5,
                      "incident_id": "INC-SYNTHETIC-1",
                      "input": {
                        "text": "차아염소산나트륨 탱크에서 누출이 있고 옆 저장고에는 염산이 있습니다.",
                        "type": "VOICE_TRANSCRIPT"
                      },
                      "request_id": "REQ-DEMO-1"
                    },
                    "invalidated_confirmation_ids": [
                      "CNF-DEMO-INCIDENT"
                    ],
                    "revision": 4
                  }
                },
                "conflict": {
                  "summary": "합성 예시: 새 근거 상충 보고",
                  "value": {
                    "analysis": {
                      "confirmed_facility_substance": {
                        "cas_number": "7647-01-0",
                        "confirmation_basis": "CONTAINER_LABEL",
                        "confirmation_id": "CNF-DEMO-FACILITY",
                        "observed_at": "2026-09-01T00:00:00+09:00",
                        "presence_status": "CONFIRMED_PRESENT",
                        "role": "FACILITY"
                      },
                      "confirmed_incident_substance": {
                        "cas_number": "7681-52-9",
                        "confirmation_basis": "CONTAINER_LABEL",
                        "confirmation_id": "CNF-DEMO-INCIDENT",
                        "observed_at": "2026-09-01T00:00:00+09:00",
                        "presence_status": "CONFIRMED_PRESENT",
                        "role": "INCIDENT"
                      },
                      "evidence_top_k": 5,
                      "incident_id": "INC-SYNTHETIC-1",
                      "input": {
                        "text": "차아염소산나트륨 탱크에서 누출이 있고 옆 저장고에는 염산이 있습니다.",
                        "type": "VOICE_TRANSCRIPT"
                      },
                      "request_id": "REQ-DEMO-1"
                    },
                    "reported_evidence_conflict": true,
                    "revision": 4
                  }
                },
                "invalid_cas": {
                  "summary": "오류 예시: CAS checksum 오류(422)",
                  "value": {
                    "analysis": {
                      "confirmed_incident_substance": {
                        "cas_number": "7681-52-0",
                        "confirmation_basis": "CONTAINER_LABEL",
                        "confirmation_id": "CNF-DEMO-INCIDENT",
                        "observed_at": "2026-09-01T00:00:00+09:00",
                        "presence_status": "CONFIRMED_PRESENT",
                        "role": "INCIDENT"
                      },
                      "evidence_top_k": 5,
                      "incident_id": "INC-SYNTHETIC-1",
                      "input": {
                        "text": "차아염소산나트륨 탱크에서 누출이 있고 옆 저장고에는 염산이 있습니다.",
                        "type": "VOICE_TRANSCRIPT"
                      },
                      "request_id": "REQ-DEMO-1"
                    },
                    "revision": 2
                  }
                },
                "no_evidence": {
                  "summary": "합성 예시: 인덱스의 해당 CAS 근거가 없을 수 있음",
                  "value": {
                    "analysis": {
                      "confirmed_incident_substance": {
                        "cas_number": "7732-18-5",
                        "confirmation_basis": "CONTAINER_LABEL",
                        "confirmation_id": "CNF-DEMO-INCIDENT",
                        "observed_at": "2026-09-01T00:00:00+09:00",
                        "presence_status": "CONFIRMED_PRESENT",
                        "role": "INCIDENT"
                      },
                      "evidence_top_k": 5,
                      "incident_id": "INC-SYNTHETIC-1",
                      "input": {
                        "text": "미상 제품의 정보 확인 요청",
                        "type": "VOICE_TRANSCRIPT"
                      },
                      "request_id": "REQ-DEMO-1"
                    },
                    "revision": 1
                  }
                },
                "one_confirmed": {
                  "summary": "합성 예시: 사고물질만 확인",
                  "value": {
                    "analysis": {
                      "confirmed_incident_substance": {
                        "cas_number": "7681-52-9",
                        "confirmation_basis": "CONTAINER_LABEL",
                        "confirmation_id": "CNF-DEMO-INCIDENT",
                        "observed_at": "2026-09-01T00:00:00+09:00",
                        "presence_status": "CONFIRMED_PRESENT",
                        "role": "INCIDENT"
                      },
                      "evidence_top_k": 5,
                      "incident_id": "INC-SYNTHETIC-1",
                      "input": {
                        "text": "차아염소산나트륨 탱크에서 누출이 있고 옆 저장고에는 염산이 있습니다.",
                        "type": "VOICE_TRANSCRIPT"
                      },
                      "request_id": "REQ-DEMO-1"
                    },
                    "revision": 2
                  }
                },
                "unconfirmed": {
                  "summary": "합성 예시: 미확인",
                  "value": {
                    "analysis": {
                      "evidence_top_k": 5,
                      "incident_id": "INC-SYNTHETIC-1",
                      "input": {
                        "text": "차아염소산나트륨 탱크에서 누출이 있고 옆 저장고에는 염산이 있습니다.",
                        "type": "VOICE_TRANSCRIPT"
                      },
                      "request_id": "REQ-DEMO-1"
                    },
                    "revision": 1
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/BriefRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "검증된 initial/final BriefResponse를 담은 SSE. 각 이벤트는 전체 snapshot이며 합쳐서 누적하지 않습니다."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "백엔드 인증 실패"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "요청 계약 위반"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "출력 안전 검증 또는 내부 오류"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "인증 또는 모델 artifact 준비 실패"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "검증된 확인 카드와 후속 결과를 SSE로 순차 반환",
        "tags": [
          "action-brief"
        ]
      }
    },
...
  }
}